created service for createNotification
