+++ NotificationApi/NotificationApi.DAL/NotificationApi.DAL.csproj
@@ -16,6 +16,7 @@
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="3.1.9" />
     <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="3.1.1" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+    <PackageReference Include="Notify" Version="2.9.0" />
     <PackageReference Include="SonarAnalyzer.CSharp" Version="8.14.0.22654">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
@@ -28,6 +29,7 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\NotificationApi.Common\NotificationApi.Common.csproj" />
+    <ProjectReference Include="..\NotificationApi.Contract\NotificationApi.Contract.csproj" />
     <ProjectReference Include="..\NotificationApi.Domain\NotificationApi.Domain.csproj" />
   </ItemGroup>
 </Project>
