+++ NotificationApi/NotificationApi.UnitTests/Controller/Notification/GetTemplateByNotificationTypeTests.cs
@@ -39,7 +39,7 @@ namespace NotificationApi.UnitTests.Controller.Notification
                 _mocker.Mock<IQueryHandler>().Setup(x => x.Handle<GetTemplateByNotificationTypeQuery, Template>(It.Is<GetTemplateByNotificationTypeQuery>(y => y.NotificationType == notificationType))).ReturnsAsync(template);
 
                 // Act
-                var result = await _sut.GetTemplateByNotificationTypeAsync((int)notificationType);
+                var result = await _sut.GetTemplateByNotificationType((int)notificationType);
 
                 // Assert
                 _mocker.Mock<IQueryHandler>().Verify(x => x.Handle<GetTemplateByNotificationTypeQuery, Template>(It.Is<GetTemplateByNotificationTypeQuery>(y => y.NotificationType == notificationType)), Times.Once);
@@ -62,7 +62,7 @@ namespace NotificationApi.UnitTests.Controller.Notification
             // Arrange
 
             // Act / Assert
-            var exception = Assert.ThrowsAsync<BadRequestException>(() => _sut.GetTemplateByNotificationTypeAsync(100000));
+            var exception = Assert.ThrowsAsync<BadRequestException>(() => _sut.GetTemplateByNotificationType(100000));
             _mocker.Mock<IQueryHandler>().Verify(x => x.Handle<GetTemplateByNotificationTypeQuery, Template>(It.IsAny<GetTemplateByNotificationTypeQuery>()), Times.Once);
             exception.Message.Should().Be($"Invalid notificationType: 100000");
         }
