+++ NotificationApi/NotificationApi.sln
@@ -21,8 +21,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NotificationApi.Common", "N
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{4078E3A9-2AB6-4EF8-8904-B26E83CCA081}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NotificationApi.Services", "NotificationApi.Services\NotificationApi.Services.csproj", "{F421A547-B56F-439F-9E84-DAA1E44980FE}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NotificationApi.AcceptanceTests", "NotificationApi.AcceptanceTests\NotificationApi.AcceptanceTests.csproj", "{B0F53513-880B-4409-BD93-AE3B668A6FD1}"
 EndProject
 Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "docker-compose", "docker-compose.dcproj", "{1AA310C5-D155-44E7-A18D-4D2564868DE0}"
@@ -69,10 +67,6 @@ Global
 		{34D8035B-3FA6-4C55-B5DB-B62FB3D4EF91}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{34D8035B-3FA6-4C55-B5DB-B62FB3D4EF91}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{34D8035B-3FA6-4C55-B5DB-B62FB3D4EF91}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F421A547-B56F-439F-9E84-DAA1E44980FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F421A547-B56F-439F-9E84-DAA1E44980FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F421A547-B56F-439F-9E84-DAA1E44980FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F421A547-B56F-439F-9E84-DAA1E44980FE}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B0F53513-880B-4409-BD93-AE3B668A6FD1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B0F53513-880B-4409-BD93-AE3B668A6FD1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B0F53513-880B-4409-BD93-AE3B668A6FD1}.Release|Any CPU.ActiveCfg = Release|Any CPU
